refactor(reader-summary): split quality dashboard responsibilities
